Sounds Correction & Wheels Collision
Corrected the sound "colliding" with checkpoints (sounds are bind to a collision channel to occlude them, it was bind to visibility, mening that when the camera is blocked by something, it would cut the sound. Now binded to Vehicle collision type). Also added collision filtering for the wheels, so they no longer collides with walls, only with the road.